Add files via upload
Slicing the text meaning to read only from one specific character to the end value we have given.
SO if we say message.slice(0,180) it will print out the message from the first character all away to character number 179.
NOTE: First character of an string has an index of 0. 
In the example we have given the slice function first parameter of 0, which means start from the first character 
and the second parameter we give to the slice function gives an index to stop, without printing that character. 
So if we wanted to print out for example in the middle of the text we can simply say message.slice(50.180) and etc. instead of the 50.